+++ lib/router.dart
@@ -1,7 +1,8 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flex_storefront/account/account_page.dart';
-import 'package:flex_storefront/account/pages/address_page.dart';
-import 'package:flex_storefront/account/pages/new_address_page.dart';
+import 'package:flex_storefront/account/pages/my_account_page.dart';
+import 'package:flex_storefront/account/pages/addresses/address_page.dart';
+import 'package:flex_storefront/account/pages/addresses/new_address_page.dart';
 import 'package:flex_storefront/analytics/apis/analytics_api.dart';
 import 'package:flex_storefront/analytics/models/analytics_events.dart';
 import 'package:flex_storefront/auth/auth_repository.dart';
@@ -99,18 +100,29 @@ class AppRouter extends _$AppRouter {
               ],
             ),
             AutoRoute(page: CartRoute.page, path: 'cart'),
-            AutoRoute(page: AccountRoute.page, path: 'account'),
+            AutoRoute(
+              page: AccountRoute.page,
+              path: 'account',
+              children: [
+                AutoRoute(
+                  page: MyAccountRoute.page,
+                  path: 'me',
+                  initial: true,
+                ),
+                AutoRoute(
+                  page: AddressRoute.page,
+                  path: 'address',
+                  guards: [AuthGuard()],
+                ),
+              ],
+            ),
           ],
-        ),
-        AutoRoute(
-          page: AddressRoute.page,
-          path: '/address',
-          fullscreenDialog: true,
         ),
         AutoRoute(
           page: NewAddressRoute.page,
           path: '/address/new',
-          // fullscreenDialog: true,
+          fullscreenDialog: true,
+          guards: [AuthGuard()],
         ),
         AutoRoute(
           page: FilterRoute.page,
